+++ CHANGELOG.md
@@ -1,3 +1,13 @@
+## [1.0.2]
+- Now when the TextField layout changes, the focus area can follow
+- New parameters:
+    - bool showFocusArea
+        Display a translucent red box to show the focus area, it's for debug.
+    - bool keepFocusStyle
+        The TextField lost focus but keep to show the focus style. 
+    - bool Function(Widget widget) onLostFocus 
+        Return true to allow lost focus.
+
 ## [1.0.1+2] - Updated README.md
 
 ## [1.0.1+1] - Updated README.md